Update the criteria to the new ones
